Inline mul.rs back into montgomery_backend.rs
Pure structural revert of the file split — the body of
generate_mul_impl is byte-for-byte identical. Keeps the mul codegen
colocated with the rest of the SmallFpConfig backend generation,
which makes the PR easier to review and keeps the is_babybear /
is_koalabear prime detection (also used by backend_impl to pin
k_bits=32) visible alongside its mul-side counterpart.

No behavior change.
